==> KCDC/HLA_imputataion/HLAtping_matching/02.Donor_recipient_HLAtype_matching_count.ipynb <==
## recipient Donor HLA typing matching count
## 20211215

# R 에서 작업후 가지고 오기
# [1] "ref"         "KR_KBA_ID"   "KR_OriID"    "KR_A.1"      "KR_A_B.1"    "KR_B.1"      "KR_B_B.1"    "KR_C.1"      "KR_C_B.1"   
#[10] "KR_DRB1.1"   "KR_DRB1_B.1" "KR_DPA1.1"   "KR_DPA1_B.1" "KR_DPB1.1"   "KR_DPB1_B.1" "KR_DQA1.1"   "KR_DQA1_B.1" "KR_DQB1.1"  
#[19] "KR_DQB1_B.1" "KD_KBA_ID"   "KD_OriID"    "KD_A.1"      "KD_A_B.1"    "KD_B.1"      "KD_B_B.1"    "KD_C.1"      "KD_C_B.1"   
#[28] "KD_DRB1.1"   "KD_DRB1_B.1" "KD_DPA1.1"   "KD_DPA1_B.1" "KD_DPB1.1"   "KD_DPB1_B.1" "KD_DQA1.1"   "KD_DQA1_B.1" "KD_DQB1.1"  
#[37] "KD_DQB1_B.1"

wdir = "/Users/ksmpooh/Desktop/KCDC/HLAimputation/03.HLA_matching/"
#fileIn = wdir + "2019_Kidney_HLAmathing_table_2digit_changeNGStyping.txt"
fileIn = wdir + "2019_Kidney_HLAmathing_table_4digit_changeNGStyping.txt"
fileOut = fileIn.replace(".txt","_score_A.B.DRB1.txt")



def filein(datain):
    a = open(datain,'r')
    return [s.replace("\n","") for s in a]

def fileout(datain,path):
    a = open(path,"w")
    for i in datain:
        a.write('\t'.join(str(s) for s in i)+'\n')
    a.close()
    return a

def match_process(ori,gene):
    df = ori
    #print(df)
    #for i in ["match","wrong","empty"]:
    #    df[0].append(i)
    #print(df) 
    df[0].append(gene+"_match")
    for index,i in enumerate(df[1:]):
        #print(index,i)
        match = 0
        wrong = 0
        empty = 0
        sm = i[0:1+1]
        ngs = i[2:3+1]
        #print(sm)
        #print(ngs)
        if (sm.count("NA") == 1) | (ngs.count("NA") == 1):
            empty = 1
        elif (sm.count("NA") == 2) | (ngs.count("NA") ==2):
            empty = 2 
        #print("count : empty %s"%str(empty))
        life = empty
        for a in sm:
            #if a == "NA":
            #    continue
            #print("a : %s"%int(a))
            if match + wrong + empty == 2:
            #    print("Break")
                break
            if a in ngs:
            #    print("match")
                match = match + 1
                ngs.remove(a)
            elif (a not in ngs) & (life == 0):
            #    print("wrong")
                wrong = wrong +  1
            elif (a not in ngs) & (life == 1):
                life = life -  1
        #print("last")
        df[index+1].append(match)
        #df[index+1].append(wrong)
        #df[index+1].append(empty)
        #print(df[index+1])
    return df
            
def gene_df(ori,a1,a2,a3,a4):
    outdf = []
    for i in ori:
        x = []
        x.append(i[int(a1)])
        x.append(i[int(a2)])
        x.append(i[int(a3)])
        x.append(i[int(a4)])
        outdf.append(x)
        #print(x)
    return outdf 




def final_processing(datain,dataout):
    df = filein(datain)
    df = [s.split("\t") for s in df]
    df = match_process(df)
    fileout(df,dataout)

genes = ["A","B","C","DRB1","DPA1","DPB1","DQA1","DQB1"]
in_df = filein(fileIn)

in_df[0:3]

index = in_df[0].split("\t")
index

out = []
in_df = [s.split("\t") for s in in_df]
for i in in_df:
    #print(i.split(",")[0:2])
    out.append(i[0:5])
out[0:3]

organ = "K"
genes = ["A","B","DRB1"]
for gene in genes:
    
    #print(gene)
    R_allele1 = index.index("%sR_%s.1"%(organ,gene))
    R_allele2 = index.index("%sR_%s.2"%(organ,gene))
    #print(IMP_allele1,IMP_allele2)
    D_allele1 = index.index("%sD_%s.1"%(organ,gene))
    D_allele2 = index.index("%sD_%s.2"%(organ,gene))
    #print(D_allele1)
    #print(D_allele2)
    #print(in_df[0:5])
    pre_df = gene_df(in_df,R_allele1,R_allele2,D_allele1,D_allele2)
    pre_df = match_process(pre_df,gene)
    #print(pre_df[0:5])    
    for i,j in enumerate(pre_df):
        for k in j:
            out[i].append(k)
    #print(pre_df)



fileout(out,fileOut)
